tests/wpt/tests/container-timing/tentative/containertiming-with-child-img.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Container Timing: observe with a node with containertiming and an image child</title>
<body>
<style>
body {
  margin: 0;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/container-timing/resources/container-timing-helpers.js"></script>
<script src="/element-timing/resources/element-timing-helpers.js"></script>
<script>
  let beforeRender;
  async_test(function (t) {
    assert_implements(window.PerformanceContainerTiming, "PerformanceContainerTiming is not implemented");
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        checkContainerEntry(entry, 'div_ct', 'img_id', beforeRender)
        checkRect(entry, [0, 100, 0, 100])
        checkContainerSize(entry, 10000);
      })
    );
    observer.observe({entryTypes: ['container']});

    // Add a div that is the container timing root
    const div = document.createElement('div');
    div.setAttribute('containertiming', 'div_ct');
    document.body.appendChild(div);

    // Add image of width equal to 100 and height equal to 100.
    const img = document.createElement('img');
    img.src = '/container-timing/resources/square100.png';
    img.setAttribute('id', 'img_id');
    div.appendChild(img);

    beforeRender = performance.now();
  }, 'Paint of the image child of container timing is reported.');
</script>

</body>
